Migrate TravisCI config to GitHub Actions
This commit removes the Travis config and replaces
it with a GitHub actions workflow, which is standard
for dxw repositories.

See:
    https://blog.travis-ci.com/2020-11-02-travis-ci-new-billing

Note that we only support versions of Node that are in long term
support, and since there is no lock file here, we use npm install,
rather than npm ci.

See:
    https://endoflife.date/nodejs